+++ app/src/main/java/com/konkuk/medicarecall/data/repository/ElderIdRepositoryImpl.kt
@@ -1,0 +1,28 @@
+package com.konkuk.medicarecall.data.repository
+
+import javax.inject.Inject
+import javax.inject.Singleton
+
+@Singleton
+class ElderIdRepositoryImpl @Inject constructor() : ElderIdRepository {
+
+    private var elderIds: MutableList<Map<String, Int>> = mutableListOf()
+
+    override fun addElderId(name: String, id: Int) {
+        this.elderIds.add(mapOf(name to id))
+    }
+
+    override fun clearElderId() {
+        elderIds.clear()
+    }
+
+    override fun getElderIds(): List<Map<String, Int>> {
+        return elderIds
+    }
+
+    override fun replaceAll(items: List<Map<String, Int>>) {
+        elderIds.clear()
+        elderIds.addAll(items)
+    }
+
+}
